Move adapters to own module
