+++ healthchain/io/__init__.py
@@ -1,7 +1,7 @@
 from .containers import DataContainer, Document, Tabular
 from .base import BaseAdapter
-from .cdaadapter import CdaAdapter
-from .cdsfhiradapter import CdsFhirAdapter
+from .adapters.cdaadapter import CdaAdapter
+from .adapters.cdsfhiradapter import CdsFhirAdapter
 
 __all__ = [
     # Containers
