+++ healthchain/io/adapters/__init__.py
@@ -1,0 +1,11 @@
+"""
+Adapters module for HealthChain.
+
+This module contains adapter implementations for converting between different
+data formats and HealthChain's internal Document representation.
+"""
+
+from .cdaadapter import CdaAdapter
+from .cdsfhiradapter import CdsFhirAdapter
+
+__all__ = ["CdaAdapter", "CdsFhirAdapter"]
